# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoRestTeach API
  description: |
    A teaching REST API built with **Go + Gin + GORM + PostgreSQL**.

    Designed to help Flutter teams learn how to work with REST endpoints and Postman.

    ## Authentication

    The API uses **JWT Bearer** authentication:
    1. Register via `POST /auth/register`
    2. Obtain tokens via `POST /auth/login`
    3. Add the header: `Authorization: Bearer <access_token>`
    4. The access token expires in **15 minutes**. Refresh it via `POST /auth/refresh`

    ## Response Format

    **Success:**
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "page": 1, "per_page": 10, "total": 42 }
    }
    ```

    **Error:**
    ```json
    {
      "success": false,
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Validation failed",
        "details": [{ "field": "email", "message": "Must be a valid email address" }]
      }
    }
    ```

    ## Error Codes

    | HTTP | Code | When |
    |------|------|------|
    | 400 | `VALIDATION_ERROR` | Invalid request body fields |
    | 400 | `BAD_REQUEST` | Malformed JSON or missing body |
    | 401 | `UNAUTHORIZED` | Missing or invalid token |
    | 401 | `TOKEN_EXPIRED` | JWT access token has expired |
    | 403 | `FORBIDDEN` | Action not allowed (e.g. editing another user's post) |
    | 404 | `NOT_FOUND` | Resource does not exist |
    | 409 | `CONFLICT` | Email already registered |
...
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Paste the `access_token` received from `/auth/login`.
        Format: `Bearer <token>`
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
              enum:
                - VALIDATION_ERROR
                - BAD_REQUEST
                - UNAUTHORIZED
                - TOKEN_EXPIRED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - FILE_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - INTERNAL_ERROR
            message:
              type: string
              example: Validation failed
            details:
              type: array
              nullable: true
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    UserPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        bio:
          type: string
          example: Flutter developer at AcidSoft
        avatar_id:
          type: string
          format: uuid
          nullable: true
          example: 660e8400-e29b-41d4-a716-446655440001
          description: Use `GET /images/{avatar_id}` to fetch the image bytes
        created_at:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authorization header is required
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Post not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: email
                  message: Must be a valid email address
                - field: password
                  message: Must be at least 8 characters long
paths:
...
  /users/me:
    get:
      tags: [users]
      summary: Get my profile
      description: |
        Returns the authenticated user's profile.

        The user identity is extracted **directly from the JWT token** — no
        need to send a user ID in the request.

        **Team task (beginner):** Add the Bearer token and view your profile.
        Decode the JWT at [jwt.io](https://jwt.io) and locate `user_id` and
        `email` in the payload.
      operationId: getMyProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserPublic'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [users]
      summary: Update my profile
      description: Updates the authenticated user's `name` and/or `bio`.
      operationId: updateMyProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: Jane Doe
                bio:
                  type: string
                  maxLength: 500
                  example: I build Flutter apps
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/avatar:
    post:
      tags: [users]
      summary: Upload avatar
      description: |
        Uploads an avatar image. The file is stored as a **bytea blob** in
        PostgreSQL — no file system needed.

        After a successful upload, `avatar_id` appears in your profile.
        Use `GET /images/{avatar_id}` to display the image.

        **Team task (intermediate):** In Postman choose
        `Body → form-data`, add a key named `avatar` with type **File**,
        and select a JPEG/PNG image. After upload, open
        `GET /images/{avatar_id}` in your browser — the image should render.

        **Allowed types:** JPEG, PNG, WebP, GIF
        **Max size:** 5 MB
      operationId: uploadAvatar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, WebP, or GIF — max 5 MB)
      responses:
        '200':
          description: Avatar uploaded; updated profile returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserPublic'
        '400':
          description: Field `avatar` is missing
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File exceeds the 5 MB limit
          content:
            application/json:
              example:
                success: false
                error:
                  code: FILE_TOO_LARGE
                  message: File exceeds maximum allowed size of 5MB
        '415':
          description: Unsupported file type
          content:
            application/json:
              example:
                success: false
                error:
                  code: UNSUPPORTED_MEDIA_TYPE
                  message: Only JPEG, PNG, WebP and GIF images are allowed
  /users/{id}:
    get:
      tags: [users]
      summary: Get user by ID
      description: Returns the public profile of any user identified by their UUID.
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserPublic'
        '400':
          description: Invalid UUID format
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── POSTS ──────────────────────────────────────────────────────────────────
...